Authorization with Pundit gem

+++ app/controllers/contacts_controller.rb
@@ -28,9 +28,12 @@ class ContactsController < ApplicationController
   end
 
   def edit
+    authorize @contact
   end
 
   def update
+    authorize @contact
+
     if @contact.update(contact_params)
       flash[:success] = "Contact was successfully updated"
       redirect_to contacts_path(previous_query_string)
@@ -40,6 +43,8 @@ class ContactsController < ApplicationController
   end
 
   def destroy
+    authorize @contact
+
     if @contact.destroy
       flash[:success] = "Contact was successfully deleted"      
     else
